Exit earlier if dependency missing
If there's a dependency missing, I don't think there's any reason to create functions and do a bunch of operations before reporting it with a crash. The CPU time and the memory footprint are minor, but it's just good practice if the crash happens before all the functions and other things are defined. Also, I think putting a `[OPTIONS]` indicator in the help screen would be nice. Many programs have it (for example, kitty, python, gcc) and it makes things clearer (where to put the flags, where to put args). Explicit is better than implicit, after all.

It's also conventional to use hyphens or underscores instead of spaces in indicators like `search-query`.

These are just suggestions. Please feel free to edit as you see fit, or close the PR if inappropriate.