feat: store and show video view dates
This commit modifies the storage of view history to include the date and
time a video was viewed. Date format is "YYYY-MM-DD HH:MM". The view
date is also displayed when loading videos from history using `-H`.

Important note: as opposed to other fields, the view date will always be
displayed in full, regardless of the terminal width. This should
probably change to avoid display issues with small terminals.

The change is backwards compatible with existing history files (the
"Viewed" column will simply not be displayed when missing).