feat: allow removing views earlier than date/interval
This commit introduces a `-X` (or `--clear-before`) option to the script,
which allows clearing the history only of videos viewed earlier than the
provided date.

The option accepts a string in any format accepted by the `date`
command's `-d` or `--date` option. This includes actual dates, or
intervals such as '2 days ago' and '24 hours ago'.

This feature is useful for automatic executions (e.g. through `cron` or
a `systemd` timer), where users might wish to retain recent history.