fix handling of curl's failure
curl's exit code will indicate failure even if it
fails in fetching only one of the thumbnails. therefore
to detect that it's failure was due to an unavailable option
(in this case the -Z option), we need to make sure it's exit
code is 2, before trying to run curl again.