fix: make instance fetching fail elegantly
Code now fails, with proper error messages, if cURL encounters errors
while fetching `instance_list`.
This is necessary so that the code doesn't simply write out an empty
`instance_list`. In the worst case scenario, it could even replace a
previously usable `instance_list` with an empty one!

Additionally, a nice status message informs the user what is going on.